+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Any case can now be used for commands.
 
+### Fixed
+
+- Navigating back through slides is now aware of `VIDEO` and `AUDIO` commands.
+
 ## [1.0.0] (2023-01-11)
 
 Initial release.